+++ lib/new_relic/agent/configuration/default_source.rb
@@ -4,6 +4,7 @@
 
 require 'forwardable'
 require_relative '../../constants'
+require_relative '../instrumentation/active_support_subscriber'
 
 module NewRelic
   module Agent
@@ -35,15 +36,6 @@ module NewRelic
       end
 
       class DefaultSource
-        BOOLEAN_MAP = {
-          'true' => true,
-          'yes' => true,
-          'on' => true,
-          'false' => false,
-          'no' => false,
-          'off' => false
-        }.freeze
-
         attr_reader :defaults
 
         extend Forwardable
@@ -71,12 +63,6 @@ module NewRelic
 
         def self.allowlist_for(key)
           value_from_defaults(key, :allowlist)
-        end
-
-        def self.boolean_for(key, value)
-          string_value = (value.respond_to?(:call) ? value.call : value).to_s
-
-          BOOLEAN_MAP.fetch(string_value, nil)
         end
 
         def self.default_for(key)
@@ -213,55 +199,16 @@ module NewRelic
           end
         end
 
-        def self.convert_to_regexp_list(raw_value)
-          value_list = convert_to_list(raw_value)
-          value_list.map do |value|
-            /#{value}/
-          end
+        def self.convert_to_regexp_list(string_array)
+          string_array.map { |value| /#{value}/ }
         end
 
-        def self.convert_to_list(value)
-          case value
-          when String
-            value.split(/\s*,\s*/)
-          when Array
-            value
-          else
-            raise ArgumentError.new("Config value '#{value}' couldn't be turned into a list.")
-          end
-        end
+        def self.convert_to_constant_list(string_array)
+          return string_array if string_array.empty?
 
-        def self.convert_to_hash(value)
-          return value if value.is_a?(Hash)
-
-          if value.is_a?(String)
-            return value.split(',').each_with_object({}) do |item, hash|
-              key, value = item.split('=')
-              hash[key] = value
-            end
-          end
-
-          raise ArgumentError.new(
-            "Config value '#{value}' of " \
-            "class #{value.class} couldn't be turned into a Hash."
-          )
-        end
-
-        SEMICOLON = ';'.freeze
-        def self.convert_to_list_on_semicolon(value)
-          case value
-          when Array then value
-          when String then value.split(SEMICOLON)
-          else NewRelic::EMPTY_ARRAY
-          end
-        end
-
-        def self.convert_to_constant_list(raw_value)
-          return NewRelic::EMPTY_ARRAY if raw_value.nil? || raw_value.empty?
-
-          constants = convert_to_list(raw_value).map! do |class_name|
+          constants = string_array.map! do |class_name|
             const = ::NewRelic::LanguageSupport.constantize(class_name)
-            NewRelic::Agent.logger.warn("Ignoring invalid constant '#{class_name}' in #{raw_value}") unless const
+            NewRelic::Agent.logger.warn("Ignoring invalid constant '#{class_name}' in #{string_array}") unless const
             const
           end
           constants.compact!
@@ -347,7 +294,8 @@ module NewRelic
           :public => true,
           :type => String,
           :allowed_from_server => false,
-          :transform => DefaultSource.method(:convert_to_list_on_semicolon),
+          :transform => proc { |v| v.is_a?(String) ? v.split(';') : v },
+          :transformed_type => Array,
           :description => 'Specify the [application name](/docs/apm/new-relic-apm/installation-configuration/name-your-application) used to aggregate data in the New Relic UI. To report data to [multiple apps at the same time](/docs/apm/new-relic-apm/installation-configuration/using-multiple-names-app), specify a list of names separated by a semicolon `;`. For example, `MyApp` or `MyStagingApp;Instance1`.'
         },
         :license_key => {
@@ -504,6 +452,7 @@ module NewRelic
           :public => true,
           :type => String,
           :allowed_from_server => false,
+          :transformed_type => Hash, # NOTE: :labels is a special case and transformed in manager.rb without a :transform key
           :description => 'A dictionary of [label names](/docs/data-analysis/user-interface-functions/labels-categories-organize-your-apps-servers) and values that will be applied to the data sent from this agent. May also be expressed as a semicolon-delimited `;` string of colon-separated `:` pairs. For example, `Server:One;Data Center:Primary`.'
         },
         :log_file_name => {
@@ -848,7 +797,6 @@ module NewRelic
           :default => {},
           :public => true,
           :type => Hash,
-          :transform => DefaultSource.method(:convert_to_hash),
           :allowed_from_server => false,
           :description => 'A hash with key/value pairs to add as custom attributes to all log events forwarded to New Relic. If sending using an environment variable, the value must be formatted like: "key1=value1,key2=value2"'
         },
@@ -863,7 +811,6 @@ module NewRelic
           :default => [],
           :public => true,
           :type => Array,
-          :transform => DefaultSource.method(:convert_to_list),
           :allowed_from_server => false,
           :description => 'A case-insensitive array or comma-delimited string containing the labels to exclude from log records.'
         },
@@ -909,7 +856,6 @@ module NewRelic
           :public => true,
           :type => Array,
           :allowed_from_server => false,
-          :transform => DefaultSource.method(:convert_to_list),
           :description => 'Prefix of attributes to exclude from all destinations. Allows `*` as wildcard at end.'
         },
         :'attributes.include' => {
@@ -917,7 +863,6 @@ module NewRelic
           :public => true,
           :type => Array,
           :allowed_from_server => false,
-          :transform => DefaultSource.method(:convert_to_list),
           :description => 'Prefix of attributes to include in all destinations. Allows `*` as wildcard at end.'
         },
         :'browser_monitoring.attributes.enabled' => {
@@ -932,7 +877,6 @@ module NewRelic
           :public => true,
           :type => Array,
           :allowed_from_server => false,
-          :transform => DefaultSource.method(:convert_to_list),
           :description => 'Prefix of attributes to exclude from browser monitoring. Allows `*` as wildcard at end.'
         },
         :'browser_monitoring.attributes.include' => {
@@ -940,7 +884,6 @@ module NewRelic
           :public => true,
           :type => Array,
           :allowed_from_server => false,
-          :transform => DefaultSource.method(:convert_to_list),
           :description => 'Prefix of attributes to include in browser monitoring. Allows `*` as wildcard at end.'
         },
         :'error_collector.attributes.enabled' => {
@@ -955,7 +898,6 @@ module NewRelic
           :public => true,
           :type => Array,
           :allowed_from_server => false,
-          :transform => DefaultSource.method(:convert_to_list),
           :description => 'Prefix of attributes to exclude from error collection. Allows `*` as wildcard at end.'
         },
         :'error_collector.attributes.include' => {
@@ -963,7 +905,6 @@ module NewRelic
           :public => true,
           :type => Array,
           :allowed_from_server => false,
-          :transform => DefaultSource.method(:convert_to_list),
           :description => 'Prefix of attributes to include in error collection. Allows `*` as wildcard at end.'
         },
         :'span_events.attributes.enabled' => {
@@ -978,7 +919,6 @@ module NewRelic
           :public => true,
           :type => Array,
           :allowed_from_server => false,
-          :transform => DefaultSource.method(:convert_to_list),
           :description => 'Prefix of attributes to exclude from span events. Allows `*` as wildcard at end.'
         },
         :'span_events.attributes.include' => {
@@ -986,7 +926,6 @@ module NewRelic
           :public => true,
           :type => Array,
           :allowed_from_server => false,
-          :transform => DefaultSource.method(:convert_to_list),
           :description => 'Prefix of attributes to include on span events. Allows `*` as wildcard at end.'
         },
         :'transaction_events.attributes.enabled' => {
@@ -1001,7 +940,6 @@ module NewRelic
           :public => true,
           :type => Array,
           :allowed_from_server => false,
-          :transform => DefaultSource.method(:convert_to_list),
           :description => 'Prefix of attributes to exclude from transaction events. Allows `*` as wildcard at end.'
         },
         :'transaction_events.attributes.include' => {
@@ -1009,7 +947,6 @@ module NewRelic
           :public => true,
           :type => Array,
           :allowed_from_server => false,
-          :transform => DefaultSource.method(:convert_to_list),
           :description => 'Prefix of attributes to include in transaction events. Allows `*` as wildcard at end.'
         },
         :'transaction_segments.attributes.enabled' => {
@@ -1024,7 +961,6 @@ module NewRelic
           :public => true,
           :type => Array,
           :allowed_from_server => false,
-          :transform => DefaultSource.method(:convert_to_list),
           :description => 'Prefix of attributes to exclude from transaction segments. Allows `*` as wildcard at end.'
         },
         :'transaction_segments.attributes.include' => {
@@ -1032,7 +968,6 @@ module NewRelic
           :public => true,
           :type => Array,
           :allowed_from_server => false,
-          :transform => DefaultSource.method(:convert_to_list),
           :description => 'Prefix of attributes to include on transaction segments. Allows `*` as wildcard at end.'
         },
         :'transaction_tracer.attributes.enabled' => {
@@ -1047,7 +982,6 @@ module NewRelic
           :public => true,
           :type => Array,
           :allowed_from_server => false,
-          :transform => DefaultSource.method(:convert_to_list),
           :description => 'Prefix of attributes to exclude from transaction traces. Allows `*` as wildcard at end.'
         },
         :'transaction_tracer.attributes.include' => {
@@ -1055,7 +989,6 @@ module NewRelic
           :public => true,
           :type => Array,
           :allowed_from_server => false,
-          :transform => DefaultSource.method(:convert_to_list),
           :description => 'Prefix of attributes to include in transaction traces. Allows `*` as wildcard at end.'
         },
         # Audit log
@@ -1072,6 +1005,7 @@ module NewRelic
           :type => Array,
           :allowed_from_server => false,
           :transform => DefaultSource.method(:convert_to_regexp_list),
+          :transformed_type => Array,
           :description => 'List of allowed endpoints to include in audit log.'
         },
         :'audit_log.path' => {
@@ -1141,6 +1075,7 @@ module NewRelic
           :type => Array,
           :allowed_from_server => false,
           :transform => proc { |arr| NewRelic::Agent.add_automatic_method_tracers(arr) },
+          :transformed_type => Array,
           :description => <<~DESCRIPTION
             An array of `CLASS#METHOD` (for instance methods) and/or `CLASS.METHOD` (for class methods) strings representing Ruby methods that the agent can automatically add custom instrumentation to. This doesn't require any modifications of the source code that defines the methods.
 
@@ -1458,7 +1393,6 @@ module NewRelic
           :public => true,
           :type => Array,
           :allowed_from_server => false,
-          :transform => DefaultSource.method(:convert_to_list),
           :description => 'Ordinarily the agent reports dyno names with a trailing dot and process ID (for example, `worker.3`). You can remove this trailing data by specifying the prefixes you want to report without trailing data (for example, `worker`).'
         },
         # Infinite tracing
@@ -1482,6 +1416,24 @@ module NewRelic
           :description => 'Configures the TCP/IP port for the trace observer Host'
         },
         # Instrumentation
+        :'instrumentation.active_support_notifications.active_support_events' => {
+          :default => NewRelic::Agent::Instrumentation::ActiveSupportSubscriber::EVENT_NAME_TO_METHOD_NAME.keys,
+          :public => true,
+          :type => Array,
+          :allowed_from_server => false,
+          :description => <<~ACTIVE_SUPPORT_EVENTS.chomp.tr("\n", ' ')
+            An allowlist array of Active Support notifications events specific to the Active Support library
+            itself that the agent should subscribe to. The Active Support library specific events focus primarily
+            on caching. Any event name not included in this list will be ignored by the agent. Provide complete event
+            names such as 'cache_fetch_hit.active_support'. Do not provide asterisks or regex patterns, and do not
+            escape any characters with backslashes.
+
+            For a complete list of all possible Active Support event names, see the
+            [list of caching names](https://edgeguides.rubyonrails.org/active_support_instrumentation.html#active-support-caching)
+            and the [list of messages names](https://edgeguides.rubyonrails.org/active_support_instrumentation.html#active-support-messages)
+            from the official Rails documentation.
+          ACTIVE_SUPPORT_EVENTS
+        },
         :'instrumentation.active_support_broadcast_logger' => {
           :default => instrumentation_value_from_boolean(:'application_logging.enabled'),
           :documentation_default => 'auto',
@@ -1674,6 +1626,7 @@ module NewRelic
           :type => Array,
           :allowed_from_server => false,
           :transform => DefaultSource.method(:convert_to_regexp_list),
+          :transformed_type => Array,
           :description => %Q(Specifies a list of hostname patterns separated by commas that will match gRPC hostnames that traffic is to be ignored by New Relic for. New Relic's gRPC client instrumentation will ignore traffic streamed to a host matching any of these patterns, and New Relic's gRPC server instrumentation will ignore traffic for a server running on a host whose hostname matches any of these patterns. By default, no traffic is ignored when gRPC instrumentation is itself enabled. For example, `"private.com$,exception.*"`)
         },
         :'instrumentation.grpc_server' => {
@@ -1891,7 +1844,6 @@ module NewRelic
           type: Array,
           dynamic_name: true,
           allowed_from_server: false,
-          :transform => DefaultSource.method(:convert_to_list),
           :description => <<~DESCRIPTION
             An array of strings to specify which keys inside a Stripe event's `user_data` hash should be reported
             to New Relic. Each string in this array will be turned into a regular expression via `Regexp.new` to
@@ -1904,7 +1856,6 @@ module NewRelic
           type: Array,
           dynamic_name: true,
           allowed_from_server: false,
-          :transform => DefaultSource.method(:convert_to_list),
           :description => <<~DESCRIPTION
             An array of strings to specify which keys and/or values inside a Stripe event's `user_data` hash should
             \tnot be reported to New Relic. Each string in this array will be turned into a regular expression via
@@ -2022,6 +1973,7 @@ module NewRelic
           :type => Array,
           :allowed_from_server => false,
           :transform => DefaultSource.method(:convert_to_regexp_list),
+          :transformed_type => Array,
           :description => 'Specify an Array of Rake tasks to automatically instrument. ' \
           'This configuration option converts the Array to a RegEx list. If you\'d like ' \
           'to allow all tasks by default, use `rake.tasks: [.+]`. No rake tasks will be ' \
@@ -2042,6 +1994,7 @@ module NewRelic
           :type => Array,
           :allowed_from_server => true,
           :transform => DefaultSource.method(:convert_to_regexp_list),
+          :transformed_type => Array,
           :description => 'Define transactions you want the agent to ignore, by specifying a list of patterns matching the URI you want to ignore. For more detail, see [the docs on ignoring specific transactions](/docs/agents/ruby-agent/api-guides/ignoring-specific-transactions/#config-ignoring).'
         },
         # Serverless
@@ -2051,6 +2004,7 @@ module NewRelic
           :type => Boolean,
           :allowed_from_server => false,
           :transform => proc { |bool| NewRelic::Agent::ServerlessHandler.env_var_set? || bool },
+          :transformed_type => Boolean,
           :description => 'If `true`, the agent will operate in a streamlined mode suitable for use with short-lived ' \
                           'serverless functions. NOTE: Only AWS Lambda functions are supported currently and this ' \
                           "option isn't intended for use without [New Relic's Ruby Lambda layer](https://docs.newrelic.com/docs/serverless-function-monitoring/aws-lambda-monitoring/get-started/monitoring-aws-lambda-serverless-monitoring/) offering."
@@ -2171,11 +2125,12 @@ module NewRelic
           :description => 'If true, the agent strips messages from all exceptions except those in the [allowed classes list](#strip_exception_messages-allowed_classes). Enabled automatically in [high security mode](/docs/accounts-partnerships/accounts/security/high-security).'
         },
         :'strip_exception_messages.allowed_classes' => {
-          :default => '',
+          :default => NewRelic::EMPTY_ARRAY,
           :public => true,
-          :type => String,
+          :type => Array,
           :allowed_from_server => false,
           :transform => DefaultSource.method(:convert_to_constant_list),
+          :transformed_type => Array,
           :description => 'Specify a list of exceptions you do not want the agent to strip when [strip_exception_messages](#strip_exception_messages-enabled) is `true`. Separate exceptions with a comma. For example, `"ImportantException,PreserveMessageException"`.'
         },
         # Agent Control
@@ -2731,7 +2686,6 @@ module NewRelic
           :type => Array,
           :external => true,
           :allowed_from_server => true,
-          :transform => DefaultSource.method(:convert_to_list),
           :description => 'Defines API paths the security agent should ignore in IAST scans. Accepts an array of regex patterns matching the URI to ignore. The regex pattern should find a complete match for the URL without the endpoint. For example, `[".*account.*"], [".*/\api\/v1\/.*?\/login"]`'
         },
         :'security.exclude_from_iast_scan.http_request_parameters.header' => {
@@ -2740,7 +2694,6 @@ module NewRelic
           :type => Array,
           :external => true,
           :allowed_from_server => true,
-          :transform => DefaultSource.method(:convert_to_list),
           :description => 'An array of HTTP request headers the security agent should ignore in IAST scans. The array should specify a list of patterns matching the headers to ignore.'
         },
         :'security.exclude_from_iast_scan.http_request_parameters.query' => {
@@ -2749,7 +2702,6 @@ module NewRelic
           :type => Array,
           :external => true,
           :allowed_from_server => true,
-          :transform => DefaultSource.method(:convert_to_list),
           :description => 'An array of HTTP request query parameters the security agent should ignore in IAST scans. The array should specify a list of patterns matching the HTTP request query parameters to ignore.'
         },
         :'security.exclude_from_iast_scan.http_request_parameters.body' => {
@@ -2758,7 +2710,6 @@ module NewRelic
           :type => Array,
           :external => true,
           :allowed_from_server => true,
-          :transform => DefaultSource.method(:convert_to_list),
           :description => 'An array of HTTP request body keys the security agent should ignore in IAST scans.'
         },
         :'security.exclude_from_iast_scan.iast_detection_category.insecure_settings' => {