+++ lib/new_relic/agent/configuration/manager.rb
@@ -16,6 +16,28 @@ module NewRelic
       class Manager
         DEPENDENCY_DETECTION_VALUES = %i[prepend chain unsatisfied].freeze
 
+        BOOLEAN_MAP = {
+          'true' => true,
+          'yes' => true,
+          'on' => true,
+          'false' => false,
+          'no' => false,
+          'off' => false
+        }.freeze
+
+        INSTRUMENTATION_VALUES = %w[chain prepend unsatisfied]
+        NUMERIC_TYPES = [Integer, Float]
+        STRINGLIKE_TYPES = [String, Symbol]
+
+        TYPE_COERCIONS = {Integer => {pattern: /^\d+$/, proc: proc { |s| s.to_i }},
+                          Float => {pattern: /^\d+\.\d+$/, proc: proc { |s| s.to_f }},
+                          Array => {proc: proc { |s| s.split(/\s*,\s*/) }},
+                          Hash => {proc: proc { |s| s.split(/\s*,\s*/).each_with_object({}) { |i, h| k, v = i.split(/\s*=\s*/); h[k] = v } }},
+                          NewRelic::Agent::Configuration::Boolean => {pattern: /^(?:#{BOOLEAN_MAP.keys.join('|')})$/,
+                                                                      proc: proc { |s| BOOLEAN_MAP[s] }}}.freeze
+
+        USER_CONFIG_CLASSES = [NewRelic::Agent::Configuration::EnvironmentSource, NewRelic::Agent::Configuration::YamlSource]
+
         # Defining these explicitly saves object allocations that we incur
         # if we use Forwardable and def_delegators.
         def [](key)
@@ -112,74 +134,139 @@ module NewRelic
             next unless config
 
             accessor = key.to_sym
+            next unless config.has_key?(accessor)
 
-            if config.has_key?(accessor)
-              begin
-                return evaluate_and_apply_transformations(accessor, config[accessor])
-              rescue
-                next
-              end
+            begin
+              return evaluate_and_apply_transformations(accessor, config[accessor], config_category(config.class))
+            rescue
+              next
             end
           end
 
           nil
         end
 
-        def evaluate_procs(value)
-          if value.respond_to?(:call)
-            instance_eval(&value)
-          else
-            value
-          end
+        def config_category(klass)
+          return :user if USER_CONFIG_CLASSES.include?(klass)
+          return :test if [DottedHash, Hash].include?(klass)
+          return :manual if klass == ManualSource
+
+          return :nr
         end
 
-        def evaluate_and_apply_transformations(key, value)
-          evaluated = evaluate_procs(value)
-          default = enforce_allowlist(key, evaluated)
-          return default if default
-
-          boolean = enforce_boolean(key, value)
-          evaluated = boolean if [true, false].include?(boolean)
+        def evaluate_and_apply_transformations(key, value, category)
+          evaluated = value.respond_to?(:call) ? instance_eval(&value) : value
+          evaluated = type_coerce(key, evaluated, category)
+          evaluated = enforce_allowlist(key, evaluated)
 
           apply_transformations(key, evaluated)
         end
 
-        def apply_transformations(key, value)
-          if transform = transform_from_default(key)
-            begin
-              transform.call(value)
-            rescue => e
-              NewRelic::Agent.logger.error("Error applying transformation for #{key}, pre-transform value was: #{value}.", e)
-              raise e
-            end
-          else
-            value
+        def boolean?(type, value)
+          return false unless type == NewRelic::Agent::Configuration::Boolean
+
+          value.class == TrueClass || value.class == FalseClass
+        end
+
+        # auto-instrumentation configuration params can be symbols or strings
+        # and unless we want to refactor the configuration hash to support both
+        # types, we handle the special case here
+        def instrumentation?(type, value)
+          return false unless type == String || type == Symbol
+          return true if INSTRUMENTATION_VALUES.include?(value.to_s)
+
+          false
+        end
+
+        def handle_nil_type(key, value, category)
+          return value if %i[manual test].include?(category)
+
+          # TODO: identify all config params such as :web_transactions_apdex
+          #       that can exist in the @config hash without having an entry
+          #       in the DEFAULTS hash. then warn here when a key is in play
+          #       that is not on that allowlist. for now, just permit any key
+          #       and return the value.
+          #       https://github.com/newrelic/newrelic-ruby-agent/issues/3340
+          default_without_warning(key) || value
+        end
+
+        # permit an int to be supplied for a float based param and vice versa
+        def numeric_conversion(value)
+          value.is_a?(Integer) ? value.to_f : value.round
+        end
+
+        # permit a symbol to be supplied for a string based param and vice versa
+        def string_conversion(value)
+          value.is_a?(Symbol) ? value.to_s : value.to_sym
+        end
+
+        def type_coerce(key, value, category)
+          return validate_nil(key, category) if value.nil?
+
+          type = DEFAULTS.dig(key, :type)
+          return handle_nil_type(key, value, category) unless type
+          return value if value.is_a?(type) || boolean?(type, value) || instrumentation?(type, value)
+          return numeric_conversion(value) if NUMERIC_TYPES.include?(type) && NUMERIC_TYPES.include?(value.class)
+          return string_conversion(value) if STRINGLIKE_TYPES.include?(type) && STRINGLIKE_TYPES.include?(value.class)
+
+          # convert bool to string for regex usage and bool hash lookup
+          value = value.to_s if type == Boolean
+          if value.class != String
+            return value if category == :test || likely_transformed_already?(key, value)
+
+            return default_with_warning(key, value, "Expected to receive a value of type #{type} but " \
+                                        "received #{value.class}.")
           end
+
+          pattern = TYPE_COERCIONS.dig(type, :pattern)
+          if pattern && value !~ pattern
+            return default_with_warning(key, value, "Expected to receive a value of type #{type} matching " \
+              "pattern '#{pattern}'.")
+          end
+
+          procedure = TYPE_COERCIONS.dig(type, :proc)
+          return value unless procedure
+
+          procedure.call(value)
+        end
+
+        def likely_transformed_already?(key, value)
+          DEFAULTS.dig(key, :transformed_type) == value.class
+        end
+
+        def default_with_warning(key, value, msg)
+          default = default_without_warning(key)
+          NewRelic::Agent.logger.warn "Received an invalid '#{value}' value for the '#{key}' configuration " \
+            "parameter! #{msg} Using the default value of '#{default}'."
+          default
+        end
+
+        def default_without_warning(key)
+          default = DEFAULTS.dig(key, :default)
+          default.respond_to?(:call) ? default.call : default
+        end
+
+        def validate_nil(key, category)
+          return if DEFAULTS.dig(key, :allow_nil) || category == :test # tests are free to specify nil
+          return default_without_warning(key) unless category == :user # only user supplied config raises a warning
+
+          default_with_warning(key, nil, 'Nil values are not permitted for the parameter.')
+        end
+
+        def apply_transformations(key, value)
+          return value unless transform = default_source.transform_for(key)
+
+          transform.call(value)
+        rescue => e
+          default_with_warning(key, value, "Error encountered while applying transformation: >>#{e}<<")
         end
 
         def enforce_allowlist(key, value)
-          return unless allowlist = default_source.allowlist_for(key)
-          return if allowlist.include?(value)
+          return value unless allowlist = default_source.allowlist_for(key)
+          return value if allowlist.include?(value)
 
-          default = default_source.default_for(key)
-          NewRelic::Agent.logger.warn "Invalid value '#{value}' for #{key}, applying default value of '#{default}'"
-          default
-        end
-
-        def enforce_boolean(key, value)
-          type = default_source.value_from_defaults(key, :type)
-          return unless type == Boolean
-
-          bool_value = default_source.boolean_for(key, value)
-          return bool_value unless bool_value.nil?
-
-          default = default_source.default_for(key)
-          NewRelic::Agent.logger.warn "Invalid value '#{value}' for #{key}, applying default value of '#{default}'"
-          default
-        end
-
-        def transform_from_default(key)
-          default_source.transform_for(key)
+          default_with_warning(key, value, 'Expected to receive a value found on the following list: ' \
+                               ">>#{allowlist}<<, but received '#{value}'.")
         end
 
         def default_source
@@ -193,17 +280,19 @@ module NewRelic
 
         def invoke_callbacks(direction, source)
           return unless source
+          return if source.respond_to?(:empty?) && source.empty?
 
           source.keys.each do |key|
+            next unless @callbacks.key?(key)
+
             begin
-              # we need to evaluate and apply transformations for the value to deal with procs as values
-              # this is usually done by the fetch method when accessing config, however the callbacks bypass that
-              evaluated_cache = evaluate_and_apply_transformations(key, @cache[key])
-              evaluated_source = evaluate_and_apply_transformations(key, source[key])
-            rescue
+              evaluated_source = evaluate_and_apply_transformations(key, source[key], config_category(source.class))
+            rescue => e
+              NewRelic::Agent.logger.warn("Error evaluating callback for direction '#{direction}' with key '#{key}': #{e}")
               next
             end
 
+            evaluated_cache = @cache[key]
             if evaluated_cache != evaluated_source
               @callbacks[key].each do |proc|
                 if direction == :add
@@ -252,8 +341,8 @@ module NewRelic
         end
 
         def apply_mask(hash)
-          MASK_DEFAULTS \
-            .select { |_, proc| proc.call } \
+          MASK_DEFAULTS
+            .select { |_, proc| proc.call }
             .each { |key, _| hash.delete(key) }
           hash
         end