+++ CHANGELOG.md
@@ -11,6 +11,7 @@
   Previously, Cross Application Tracing (CAT) was deprecated in favor of Distributed Tracing. CAT functionality has now been removed. The configuration option `cross_application_tracer.enabled` has been removed. Public API methods `NewRelic::Agent::External.process_request_metadata`, `NewRelic::Agent::External.get_response_metadata`, `NewRelic::Agent::Transaction::ExternalRequestSegment#process_response_metadata`, `NewRelic::Agent::Transaction::ExternalRequestSegment#get_request_metadata`, and `NewRelic::Agent::Transaction::ExternalRequestSegment#read_response_headers` have also been removed. [PR#3333](https://github.com/newrelic/newrelic-ruby-agent/pull/3333)
 
 - **Feature: Add `logger` as a dependency**
+
   The `logger` gem is now listed as a dependency of the agent to ensure continued logging functionality and support for Ruby 4.0.0 and newer versions. [PR#3293](https://github.com/newrelic/newrelic-ruby-agent/pull/3293)
 
 - **Breaking Change: Rename ActiveJob metrics**
@@ -25,19 +26,51 @@
 
 - **Breaking Change: Remove the `newrelic deployments` CLI command**
 
-    The deprecated `newrelic deployments` CLI command has been removed. To track changes and deployments in New Relic, please see our guide to [Change Tracking](https://docs.newrelic.com/docs/change-tracking/change-tracking-introduction/) for a list of available options. [PR#3299](https://github.com/newrelic/newrelic-ruby-agent/pull/3299)
+  The deprecated `newrelic deployments` CLI command has been removed. To track changes and deployments in New Relic, please see our guide to [Change Tracking](https://docs.newrelic.com/docs/change-tracking/change-tracking-introduction/) for a list of available options. [PR#3299](https://github.com/newrelic/newrelic-ruby-agent/pull/3299)
 
-- **Feature: Add argument validation for the `NewRelic::Agent#record_custom_event` API**
+- **Breaking Change: Remove the NewRelic::Agent::SqlSampler#notice_sql method**
 
-  The `NewRelic::Agent#record_custom_event` API now raises an `ArgumentError` when an invalid `event_type` is provided. A valid event type must consist only of alphanumeric characters, underscores (`_`), colons (`:`), or spaces (` `). [PR#3319](https://github.com/newrelic/newrelic-ruby-agent/pull/3319)
+  Users should call `NewRelic::Agent::Datastores.notice_sql` instead. [PR#3338](https://github.com/newrelic/newrelic-ruby-agent/pull/3338)
+
+- **Breaking Change: Remove unused arguments from various NewRelic::Agent::Datastores APIs**
+
+  The following APIs from the `NewRelic::Agent::Datastores` class have had method arguments removed:
+  * `NewRelic::Agent::Datastores.notice_sql`, previously had three positional arguments, `query`, `scoped_metric` and `elapsed`. Now, it only has `query`.
+  * `NewRelic::Agent::Datastores.notice_statement`, previously had two positional arguments `query` and `elapsed`. Now it only has `query`.
+  * `NewRelic::Agent::Datastores.wrap` requires a proc. Previously the proc received three arguments: the result of the yield, the most specific scoped metric name, and the elapsed time of the call. Now, it only receives one: the result of the yield.
+
+  The values of the removed arguments are derived from the current segment at the time of the call. [PR#3347](https://github.com/newrelic/newrelic-ruby-agent/pull/3347)
 
 - **Breaking Change: Remove experimental feature Configurable Security Policies (CSP)**
 
   The experimental feature, Configurable Security Policies (CSP), is no longer supported and has been removed. [PR#3292](https://github.com/newrelic/newrelic-ruby-agent/pull/3292)
 
+- **Feature: Add Active Support notification allowlist configuration option**
+
+  A new configuration option, `instrumentation.active_support_notifications.active_support_events`, allows users to define an allowlist of Active
+  Support notifications event names for the agent to subscribe to. By default, the agent subscribes to all [Active Support: Caching](https://guides.rubyonrails.org/active_support_instrumentation.html#active-support-caching) and [Active Support: Messages](https://guides.rubyonrails.org/active_support_instrumentation.html#active-support-messages) events. [PR#3327](https://github.com/newrelic/newrelic-ruby-agent/pull/3327)
+
+- **Feature: Use Ruby's built-in Gzip compression**
+
+  The agent now uses the built-in `Zlib.gzip` method from the Ruby standard library for compression, replacing the previous custom implementation. [PR#3332](https://github.com/newrelic/newrelic-ruby-agent/pull/3332)
+
 - **Breaking Change: Remove support for Puma versions < 3.9.0**
 
   The minimum version of Puma now supported is 3.9.0 or higher. [PR#3326](https://github.com/newrelic/newrelic-ruby-agent/pull/3326)
+
+- **Breaking Change: Improve configuration validation and coercion**
+
+  The internals used to coerce and validate the values provided for agent configuration are now more performant and more accurate.
+    * Warning messages will now be logged to the newrelic_agent.log file when nil is provided as a config value for a setting that does not support it.
+    * Integer values are permitted for Float configuration types
+    * Float values are permitted for Integer configuration types
+    * Fatal interruptions are prevented when a default value can be found to replace an invalid input value
+  [PR#3341](https://github.com/newrelic/newrelic-ruby-agent/pull/3341)
+
+- **Feature: Add argument validation for the `NewRelic::Agent#record_custom_event` API**
+
+  The `NewRelic::Agent#record_custom_event` API now raises an `ArgumentError` when an invalid `event_type` is provided. A valid event type must consist only of alphanumeric characters, underscores (`_`), colons (`:`), or spaces (` `). [PR#3319](https://github.com/newrelic/newrelic-ruby-agent/pull/3319)
+
 
 ## v9.23.0